CI & refactor
