+++ README.md
@@ -1,1 +1,5 @@
-# mocap4ros2_optitrack
+# mocap_optitrack_driver
+
+[![GitHub Action
+Status](https://github.com/MOCAP4ROS2-Project/mocap_optitrack_driver/workflows/main/badge.svg)](https://github.com/MOCAP4ROS2-Project/mocap_optitrack_driver)
+[![codecov](https://codecov.io/gh/MOCAP4ROS2-Project/mocap_optitrack_driver/main/graph/badge.svg)](https://codecov.io/gh/MOCAP4ROS2-Project/mocap_optitrack_driver)
