+++ dependency_repos.repos
@@ -1,0 +1,9 @@
+repositories:
+  mocap_msgs:
+    type: git
+    url: https://github.com/MOCAP4ROS2-Project/mocap_msgs.git
+    version: master
+  mocap:
+    type: git
+    url: https://github.com/MOCAP4ROS2-Project/mocap.git
+    version: main
