+++ mocap_optitrack_driver/launch/optitrack2.launch.py
@@ -34,7 +34,7 @@ import lifecycle_msgs.msg
 def generate_launch_description():
 
     params_file_path = os.path.join(get_package_share_directory(
-      'optitrack2_driver'), 'config', 'optitrack2_driver_params.yaml')
+      'mocap_optitrack_driver'), 'config', 'mocap_optitrack_driver_params.yaml')
 
     stdout_linebuf_envvar = SetEnvironmentVariable(
         'RCUTILS_CONSOLE_STDOUT_LINE_BUFFERED', '1')
@@ -44,9 +44,9 @@ def generate_launch_description():
     # print('')
 
     driver_node = LifecycleNode(
-        node_name='optitrack2_driver_node',
-        package='optitrack2_driver',
-        node_executable='optitrack2_driver_main',
+        node_name='mocap_optitrack_driver_node',
+        package='mocap_optitrack_driver',
+        node_executable='mocap_optitrack_driver_main',
         output='screen',
         parameters=[params_file_path],
     )
