+++ mocap_optitrack_driver/src/mocap_optitrack_driver_main.cpp
@@ -19,14 +19,14 @@
 
 #include <iostream>
 #include <memory>
-#include "optitrack2_driver/optitrack2_driver.hpp"
+#include "mocap_optitrack_driver/mocap_optitrack_driver.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 int main(int argc, char * argv[])
 {
   rclcpp::init(argc, argv);
   rclcpp::NodeOptions node_options;
-  auto node = std::make_shared<OptitrackDriverNode>();
+  auto node = std::make_shared<mocap_optitrack_driver::OptitrackDriverNode>();
   rclcpp::spin(node->get_node_base_interface());
   rclcpp::shutdown();
